pyproject.toml
[project]
name = "demo-py-telemetry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.124.4",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-exporter-otlp-proto-grpc>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-instrumentation-logging>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "pre-commit>=4.5.0",
    "structlog>=25.5.0",
]
